# api/openapi.yaml
components:
  schemas:
...
    EntryRow:
      additionalProperties: true
      properties:
        anchor_message_id:
          format: int64
          type: integer
        attachment_count:
          format: int64
          type: integer
        attachment_size:
          format: int64
          type: integer
        conversation_id:
          format: int64
          type: integer
        conversation_type:
          type: string
        counterpart_participant_id:
          format: int64
          type: integer
        deleted_from_source:
          type: boolean
        has_attachments:
          type: boolean
        key:
          type: string
        kind:
          type: string
        match:
          $ref: "#/components/schemas/MatchSummary"
        message_count:
          format: int64
          type: integer
        message_type:
          type: string
        occurred_at:
          format: date-time
          type: string
        participant_ids:
          items:
            format: int64
            type: integer
          type:
            - array
            - "null"
        participant_labels:
          items:
            type: string
          type:
            - array
            - "null"
        preview:
          type: string
        source_id:
          format: int64
          type: integer
        source_identifier:
          type: string
        source_type:
          type: string
        title:
          type: string
      required:
        - key
        - kind
        - occurred_at
        - match
        - source_id
        - source_type
        - source_identifier
        - message_type
        - conversation_type
        - title
        - preview
        - message_count
        - has_attachments
        - attachment_count
        - attachment_size
        - deleted_from_source
      type: object
    ErrorResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
    ExploreCacheUnavailableResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
        readiness:
          enum:
            - absent
            - interrupted
            - stale_schema
            - drifted
          type: string
        recovery_action:
          type: string
      required:
        - error
        - message
        - readiness
        - recovery_action
      type: object
...
    ExploreFilter:
      additionalProperties: false
      properties:
        dimension:
          enum:
            - source
            - participant
            - domain
            - message_type
            - after
            - before
            - deletion
          type: string
        values:
          items:
            type: string
          minItems: 1
          type: array
      required:
        - dimension
        - values
      type: object
    ExploreGroupDimension:
      enum:
        - source
        - participant
        - domain
        - message_type
        - kind
        - year
        - month
      type: string
...
    ExploreHTTPRequest:
      additionalProperties: false
      properties:
        candidate_snapshot_id:
          type: string
        cursor:
          type: string
        filters:
          items:
            $ref: "#/components/schemas/ExploreFilter"
          type:
            - array
            - "null"
        grouping:
          items:
            $ref: "#/components/schemas/ExploreGroupDimension"
          type:
            - array
            - "null"
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        presentation:
          enum:
            - table
            - timeline
            - files
          type: string
        query:
          type: string
        search_mode:
          enum:
            - full_text
            - semantic
            - hybrid
          type: string
        sort:
          items:
            $ref: "#/components/schemas/ExploreSort"
          type:
            - array
            - "null"
      type: object
    ExploreHTTPResponse:
      additionalProperties: true
      properties:
        cache_revision:
          type: string
        candidate_pool_saturated:
          type: boolean
        candidate_snapshot_id:
          type: string
        next_cursor:
          type: string
        rows:
          items:
            $ref: "#/components/schemas/EntryRow"
          type: array
        search_provenance:
          $ref: "#/components/schemas/SearchProvenance"
        total_count:
          format: int64
          type: integer
      required:
        - rows
        - cache_revision
        - search_provenance
      type: object
...
    ExploreSort:
      additionalProperties: false
      properties:
        direction:
          enum:
            - desc
          type: string
        field:
          enum:
            - occurred_at
          type: string
      required:
        - field
        - direction
      type: object
...
    GmailIDsResponse:
      additionalProperties: true
      properties:
        gmail_ids:
          items:
            type: string
          type:
            - array
            - "null"
      required:
        - gmail_ids
      type: object
...
    MatchSummary:
      additionalProperties: true
      properties:
        lexical_match_count:
          format: int64
          type: integer
        semantic_score:
          format: double
          type: number
        strongest_excerpt:
          type: string
      type: object
...
    SearchProvenance:
      additionalProperties: false
      properties:
        lexical_index_revision:
          type: string
        vector_generation:
          format: int64
          type: integer
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.24.0
openapi: 3.1.0
paths:
...
  /api/v1/explore:
    post:
      operationId: explore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Explore canonical archive entries
      tags:
        - Exploration
...
  /api/v1/messages/gmail-ids:
    get:
      operationId: getGmailIDsByFilter
      parameters:
        - description: Sender email/address filter
          in: query
          name: sender
          schema:
            type: string
        - description: Sender display-name filter
          in: query
          name: sender_name
          schema:
            type: string
        - description: Recipient email/address filter
          in: query
          name: recipient
          schema:
            type: string
        - description: Recipient display-name filter
          in: query
          name: recipient_name
          schema:
            type: string
        - description: Domain filter
          in: query
          name: domain
          schema:
            type: string
        - description: Label filter
          in: query
          name: label
          schema:
            type: string
        - description: Message type filter
          in: query
          name: message_type
          schema:
            type: string
        - description: Named time period
          in: query
          name: time_period
          schema:
            type: string
        - description: Time bucket granularity
          in: query
          name: time_granularity
          schema:
            type: string
        - description: Conversation ID
          in: query
          name: conversation_id
          schema:
            format: int64
            type: integer
        - description: Source ID
          in: query
          name: source_id
          schema:
            format: int64
            type: integer
        - description: Only include messages with attachments
          in: query
          name: attachments_only
          schema:
            type: boolean
        - description: Exclude deleted messages
          in: query
          name: hide_deleted
          schema:
            type: boolean
        - description: Lower date/time bound (RFC3339 or YYYY-MM-DD)
          in: query
          name: after
          schema:
            type: string
        - description: Upper date/time bound (RFC3339 or YYYY-MM-DD)
          in: query
          name: before
          schema:
            type: string
        - description: Comma-separated aggregate view names to match empty values
          in: query
          name: empty_targets
          schema:
            type: string
        - description: Zero-based row offset
          in: query
          name: offset
          schema:
            format: int64
            type: integer
        - description: Maximum number of rows to return (default and max 500; larger values are clamped)
          in: query
          name: limit
          schema:
            format: int64
            type: integer
        - description: "Sort field: date, size, or subject"
          in: query
          name: sort
          schema:
            type: string
        - description: "Sort direction: asc or desc"
          in: query
          name: direction
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailIDsResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: List Gmail message IDs matching a filter
      tags:
        - API
...